Rename three-raycast-patches to three-patches
